Split Sender and Receiver into Separate Files leaving only admin settings in plugin.
